debt: cleanup sandboxed version
